move graph walkers to paddle/common and fix reference
