+++ paddle/cinn/common/bfs_walker.h
@@ -14,59 +14,13 @@
 
 #pragma once
 
-#include <array>
-#include <functional>
-#include <queue>
-#include <unordered_set>
+#include "paddle/common/bfs_walker.h"
 
 namespace cinn {
 namespace common {
 
-// breadth-first search visitor
 template <typename NodeType>
-class BfsWalker final {
- public:
-  BfsWalker(const BfsWalker&) = delete;
-  BfsWalker(BfsWalker&&) = delete;
-
-  using NodeHandlerType = std::function<void(NodeType)>;
-  using NodesVisitorType =
-      std::function<void(NodeType, const NodeHandlerType&)>;
-
-  BfsWalker(const NodesVisitorType& VisitNextNodes)
-      : VisitNextNodes_(VisitNextNodes) {}
-
-  void operator()(NodeType node, const NodeHandlerType& NodeHandler) const {
-    std::array<NodeType, 1> nodes{node};
-    (*this)(nodes.begin(), nodes.end(), NodeHandler);
-  }
-
-  template <typename NodeIt>
-  void operator()(NodeIt begin,
-                  NodeIt end,
-                  const NodeHandlerType& NodeHandler) const {
-    std::queue<NodeType> node_queue;
-    std::unordered_set<NodeType> queued_nodes;
-    const auto& TryEnqueueNode = [&](NodeType node) {
-      if (queued_nodes.count(node) == 0) {
-        node_queue.push(node);
-        queued_nodes.insert(node);
-      }
-    };
-    for (NodeIt iter = begin; iter != end; ++iter) {
-      TryEnqueueNode(*iter);
-    }
-    while (!node_queue.empty()) {
-      NodeType node = node_queue.front();
-      node_queue.pop();
-      NodeHandler(node);
-      VisitNextNodes_(node, TryEnqueueNode);
-    }
-  }
-
- private:
-  NodesVisitorType VisitNextNodes_;
-};
+using BfsWalker = ::common::BfsWalker<NodeType>;
 
 }  // namespace common
 }  // namespace cinn
